[IMP] website_slides : add attendee status, improve invitations to course
RATIONALE

Improve and make easier the invitations on courses, whatever the enroll policy.
The goal is to increase the number of attendees easily.

PREVIOUS BEHAVIOUR

Before this commit, the attendees of a course were always enrolled.
No distinction existed. Inviting partners meant enrolling them at once, even if
they did not want to join. Also, invitations to a 'members only'-visibility
course were a problem since the invited member could not reach the course page
if not logged in. Therefore the invitation link in invitation mail was often
not usable (403 error). The invitation link did not act as a user creation link
if the partner was not linked to any user.

PURPOSE OF THIS COMMIT

After this commit, attendees to a course can either be enrolled or have pending
invitation. Invited partners have pending invitation, other ones are enrolled
by default. When creating an attendee from backend, it is set as enrolled.
Also, the invitation link is repaired and redirects the user to password
creation / course / all courses page depending on if the invited partner has a
user linked to it, if it is logged in and if it has the rights to read the
channel. Attendees also get a new member_status variable depending on their
progression in the course covering both invitation and completeness

-- ENROLLMENT

An enrolled member has the same rights as in previously existing behaviour. A
not-enrolled member will be able to see the course, but not the slides, just
like non-attendees.
    -ACL : The ir rule for slide.slide has been updated accordingly.
It also make sure invited members can access preview slides on 'members-only'
visibility channels they are invited to.

It can also self-enroll on the courses of which it is a pending member, even on
'invite'-enrollment ones, without needing to wait for manager approval. Members
are added to channel followers only once enrolled.

-- SLIDE CHANNEL PARTNER BACKEND CREATION

slide_channel_partner back-end creation is made easier. Tree view has now a
column partner_id, allowing the attendee creation, but only for partners linked
to a user, and is readonly once set to avoid status ambiguity. It also has a
member_status column. This gives a better insight of member completion.

=> In the list view, we remove the create and add share (no enroll) and invite
(enroll) buttons instead.

-- MEMBER STATUS

This new variable covers both 'completed' state (that variable is removed), and
the fact that a member is invited (pending) or has joined the course. It has 4
possible values:
- 'invite_sent' : member with pending invitation
- 'joined' : enrolled member, not started
- 'ongoing' : enrolled member, started but never finished
- 'finished' : enrolled member who completed the course
The only moment a pending member is enrolled is when joining the course directly.
The member_status is not a compute field as it is updated when progress is made,
in the tailored method _recompute_completion.

-- AUTO-PUBLISH

All the 'invite' menu wording is set to 'share' as auto-publish is implemented:
If I share a course, and it is not published on the website, then if I have the
rights and an invitation is indeed issued, the course is published on the spot.

-- VIEWS AND OTHER MAIN CHANGES

- Pivot and graph reporting views are added. A default member_status groupby too,
on all reporting views. The % of completed slides is not used as measure on pivot
view. However, it is on the graph view to compare completion of different members
easily. Avg is used as an operator, as sum of percentages does not mean much here.
- To ease the invitation process, an invite button is added at the top of the
tree view of attendees. This can be used to invite partners who do not have
linked users.
- slide_channel_partner kanban view now displays numbers of members : pending /
ongoing / finished / total as the difference can now be made between the
attendees. Corresponding filters are implemented on slide_channel_partner,
as well as course and status group_by filters.
- Quicksearch 'Tags' and 'Responsible' on slide.channel model.
- slide_channel_invite wizard now displays the invitation link for public
courses. (mail composer) It also accepts invitation no matter what user
invitation policy is set, since different 'join the course' links are sent
accordingly. Please refer to the docstrings of _get_channel_invitation_link
and slide_channel_invite methods
- On tree view of slide_channel_partner, domain is set to prevent user of
selecting a partner already in selected course, or a course containing the
partner already.
- Warning is removed from slide_channel_invite since now, both b2b and b2c
invitation policies are supported, since the invitation link is a token based
password creation email if partner has no user and course has members-only
visibility.
- Use 'course' instead of 'channel' in readable labels + heaps of small ux
improvements.
- Small scss file added to use responsive display with default color for fa
icon in button of attendees on the kanban card of courses.

--TESTS

Tests are added for the different new flows coming from this new distinction.
They test functionality, security (access rights) and model correctness.

--- Links ---
Task-2508019
COM PR - odoo/odoo#70291
UPG PR - odoo/upgrade#2572

+

[WIP] Pre-review commit : remove create from list view and add invite button.

Share = Promote your course
Invite = Invite members and enroll them
-> Owlification done

+

    [WIP] hash / review tde

    link flow
    tests adapted part 1 (todo : write tests for link)

    TODO
    adapt to the change
            channel_ALL_ids / partner_ALL_ids
    ACLS
    Tests link

( name in title )
( think about it, see notes at the end of T10 txt )

+

[WIP] tests / /invite route

TODO : tests with auth login and auth signup tokens.

Define what to do

Avoid looping in /invite route?