[IMP] web_editor: add a right margin to image inside buttons with text
Before this commit when a we-button contained both an image and a text
there was no gap between them.

After this commit a right margin is added to the image element of a
we-button whenever the button also has a text.
We cannot use the browser's built-in facilities (such as `innerText`) to
detect whether some text is present because texts contained inside
images are also returned. E.g.: the text 'LOGO' is returned from the
`header_template_slogan.svg` because it is written inside a text node of
the SVG file.

task-2359250