[IMP] crm: redefine won/lost and apply rules in PLS frequencies/views
Purpose
=======
Ribbons and buttons display + PLS won/lost frequency rules should be aligned.
 --> Redefine Won / Lost heuristic and apply it everywhere.

Specification
=============

"New heuristic":
Lost: Proba = 0 AND archived (and not in won stage)
Won: won stage AND 100% (was previously: won stage and active)
	(Won stage implies 100%.)

That means that even if a lead is archived, it is still considered as won if
he is in won stage. That also mean that we should block the edition of
probability in won stage as it should always be 100%.
In the same way, we should block edition on probability if the lead is lost.

Won / lost ribbons display rules should be aligned.

That also implies to change a bit the rules to handle won/lost frequencies

- Set to Won Stage (>WS) : Reach Won
	If probability was 0 and active was False: Leave Lost
- Move out from Won stage (<WS) : Leave Won
	If active was False and Proba = 0 -> This should never happen
- Archive: If set proba=0 or proba was 0: Reach Lost
- Restore: Leave Lost

Special cases :
< WS + Archive : "Leave Won"
	Later, if set P = 0% -> trigger "Reach Lost".
> WS + Archive : Reach Won
	If proba set to 0 -> Should never happen
< WS + Restore: Leave Won
> WS + Restore: Reach Won
	If proba was 0 : Leave Lost

This commit adds two fields on crm.lead to ease domains to display/hide
fields, buttons and ribbons:
- is_lost: computed on active=False AND proba = 0. Tracking enbaled to follow
	the state of the lead in the chatter. (when it was lost or restored)
- is_won: related to the stage_id.won_stage. Used to know if the lead is won.
	As won = in won stage.

Review the Restore vs Unarchive:
- Unarchive = set active=True
- Restore = Unarchive + force automated probability (and it's recomputation).
Restore button is displayed only on lost leads.

Probabilty is set to readonly mode when the lead is won or lost. This is to
avoid inconsistencies between the state of the lead and it's probability.

Why oh why ?
============

Before this change, the old won/lost rules where the following:
Lost: Proba = 0 OR archived
Won: won stage AND active

Which means that both rules could have coexisted. If the lead were in won stage,
active but probability = 0%. (that should not happens but was still allowed).
The objective with the new rules is to use the probability as a necessary
but not sufficient criteriat. Probability makes it impossible to be both
lost and won. But probability is not sufficient to classify a lead as won or
lost.
0% does not mean it's lost, it must also be archived.
100% does not mean it's won, it must also be in won stage.

That will avoid having automatically lost or won leads.
If Predictive lead scoring computes 0% or 100% chance of winning, the lead will
not be flagged yet as lost or won and the user will still need to take action to
correctly classify the lead.

Task-2654916
COM PR odoo/odoo#77089