[REF] mail: move 'partner_ids' field from onchange to computed
PURPOSE

Purpose of this task is to remove the _onchange_template_id method on composer
model. Split it into editable computed stored fields. It gives a better control
of value generation and avoid having to call the onchange when composer is
invoked in code.

SPECIFICATIONS

Improve composer to correctly compute fields ``partner_ids`` in compute
method instead of in an onchange method and in default_get.

Task-2088884 (MailComposer: onchange to computed stored field