[FIX] website, *: keep table of contents title and menu entries in sync
*: web_editor

The table of contents menu entries are generated automatically which
poses a problem in translation mode. The menu translation entries
are not be editable separately, but the users might be trying to.

This commit shows a notification when the user clicks on the menu
entries while in translation mode, explaining that they are generated
from the title entries.
It was initially intended to use a tooltip - but the amount of code
needed to display a tooltip without marking the DOM as modified is
needlessly complex.
Additionally, the translation widget updates the generated links of
table of content headers to keep them synchronized.
This could not be done by matching the translation id because those
are different when a style is applied to the header.

A span had to be added around navigation entries to make sure that
their translation id is distinct from the one inside the main content.
Without this, styles applied on a plain text during translation were
also appearing in the navigation menu.

task-2752391