[IMP] website_event: consider address while searching events
Currently, website users/visitors are not able to search the events based on
where they happen.

This commit improves the behavior by allowing the users/visitors to type the
address in the search box and returns the events that are happening in the
address that matches the search term. To allow this, we have utilized
`search_extra` search option and searching the matching events with `sudo`
because public users can not access the `address_search` m2o of the event.

Note that searching with sudo is just to add event ids in the domain. The final
result will be returned with non sudoed environment and so the record rules are
always respected.

TaskId-2791031