[IMP] website: add options for file uploading in forms (size + number)
Currently, we can only upload one file in a form and this file doesn't
have a size limit.

This commit adds options to the forms in order to set a maximum number
of files and the maximum file size (in MB) allowed to be uploaded in
the form. The default values are 1 file and 1 MB.

A tooltip informing the user about these limits is displayed when
hovering over the field. If he uploads too many files or if they are
too big, the form is not sent and a message is displayed.

task-2798576