[FIX] web_editor: prevent unwrapping some links
Before this commit, clicking on a social media link without an url
openned the linktool and the link was removed when the linktool
was destroyed. The links of the social media should never be
removed even if the url is empty. One reason is that it does not
update the snippet option of the dom changes and the widget
has a wrong state.

The fix prevent from unwrapping any link if it was not created before
opening the linktool for potential other cases similar to the social
snippet option.

task-2739302