[IMP] base: remove circular dependency between fields
The field ir.rule.global depends on ir.rule.groups, and its inverse field res.group.rule_groups
depends on ir.rule.global because of a domain on the field.
As the domain is only useful client-side, turn the domain into a string, so that
it is only valid client-side.