[IMP] base: multi-languages usability
Before only languages installed (aka active) are displayed to the
user. This prevents the selection of a language that does not
require translations. For example, you could specify the language
of a contact without needing the translation/multi-lang feature.

Now we have a way to display languages installed and the default
languages. The default languages refer to all languages with
an iso code of 2 letters (en, fr, es, ...)

task-2856992