[REF] mass_mailing: cleanup subscriptions update (optin / optout from lists)
PURPOSE

Overall cleaning of subscription and exclusion management code from portal.

SPECIFICATIONS

Purpose of this commit is to cleanup code that manages chosen lists from
portal and updates opt-in and opt-out accordingly.

Code is now located on mailing list model. When opting-it, new subscriptions
have to be created for public mailing lists so that email is part of the list.
When opting-out we just have to toggle the opt_out flag on subscription model.

Logged messages are also improved.

Task-2669037 (Mass Mailing: Subscription code refactoring)