[IMP] website: manage better the files uploaded in a form
Before this commit, we could observe that now that multiple files can
be uploaded at once in a form, some behaviors of the file input proved
to not be really convenient.
For example:
- clicking again on the button to choose other files replaces the
previously uploaded files by the new ones, we can therefore not upload
files one by one;

- a file cannot be deleted, unless we replace all the files (see
previous point);

- when multiple files are uploaded, only the number of files is
displayed and not their name.

This commit improves these behaviors. Now,
- clicking again on the button adds additional files instead of
replacing them (unless the number of files allowed is 1);

- the button text is modified after the uploading of files to show
better what its action will be: if only one file can be uploaded, the
text displayed is "Replace File" and if multiple files are allowed,
then it is "Add Files";

- when files are uploaded, a block containing the file name and a cross
is displayed above the button for each file. The cross allows to delete
the corresponding file.

task-2798576