[REF] mass_mailing: add a 'my' page to handle subscriptions
PURPOSE

Improve and ease subscription and exclusion list portal pages and management.

SPECIFICATIONS

In this commit a new global page is added to manage subscriptions on mailing
lists as well as block list management. It is located on ``mailing/my`` and
has the same capabilities as the unsubscribe pages, except it does not work
within a mailing's context.

Task-2150462 (Mass Mailing: Unsubscribe flow improvement)