[IMP] sale_{project,timesheet}: add task template to product
- https://nimb.ws/MLSlid + https://nimb.ws/AFv3hS
Product Form View: Add a 'Task Template' field
below the 'Project' / 'Project Template' ones
Only visible if 'Project & Task' or 'Task' are selected
for the 'Create on Order' field,
Not required, any Task can be selected as long as
it is not private (= with a project_id false),
If set, copy the selected Task when generating
a new Task on the SO confirmation.

- If all of the following conditions are met,
set the end_date of the Subscription
as the repeat_until date of the Task:
The Product is recurrent and creates a Task,
The Task Template is recurrent,
The end_date field is set on the Subscription.

- If all of the following conditions are met,
apply the Subscription's recurrence to the Task's:
The Product is recurrent and creates a Task,
There is no Task Template on the Product.

- If the Subscription is closed, set the recurring_task
field of the related Task to false.

task-2868306