[IMP] account: bank statement report improvement
Improve the bank statement report by reducing the amount of
blank lines in it, as well as adding more information about
the partner and bank account when we have it.

Also, when not using online sync the bank statement report will
now be generated at the time of posting and not validating.

Task id #2860753