[IMP] mail, sms: allow to log a message when (un)blacklisting
Add possibility to add a message when adding or removing a blacklist. This
replaces the ``action_remove_with_reason`` method. When updating active flag
this is added as a tracking note to avoid having several messages. When a
new record is created, a note is logged.

Task-2150462 (Mass Mailing: Unsubscribe flow improvement)