[IMP] survey: customize preview link design
Purpose:
--------
When sharing URLs on social medias, a link preview is usually shown.
When sharing a survey (or a survey's results), this preview now shows info
related to the survey (survey's title, description, and website), as well as
an image generated on the fly using the survey's information (title, user
and date, background image, some statistics, and "powered by odoo" logo).

Task-2842682