[IMP] digest: add app image on digest title
With a lot of digests, it is hard to know to which app it is related
(Deliveries could be about Email, Stock, "Engagement", ...), this solves the
problem by displaying the app image for each kpi.

Technical notes:
- To determine the module name, it relies on class meta-data which might
evolves in the future
- It assumes that the app icon is always located in the URL path
/<module_name>/static/description/icon.png
- Technical module don't have icons but it seems it cannot be determined
easily/efficiently, so we have added a method _compute_kpis_app_name in digests
that can be overridden to defines to which app the added kpi belong. If it is
not done, the icon won't be dispalyed thanks to an onerror attribute. But this
could decrease the score of the mail.

Task-2705154