[FIX] web_editor: allow theme blocks to specify default image processing
Image processing options (such as Filter) are not properly taken into
account if they are used inside the block customizations done by themes.

This commit recognizes the `<img>` from blocks that do not have an `src`
attribute but have a `data-original-src` instead - and use at least one
image processing option. When encountered, it applies the transformation
when the drag operation starts.

Steps to reproduce:
- Patch the "Text - Image" block:
* Rename the `src` of the `<img>` to `data-original-src`.
* Add the `data-gl-filter="1977"` parameter.
- Drop the block.
=> Filter was not taken into account.

task-2888667