[DEMO] website_coconuts: adapt demo theme to apply filters on the fly
To demo, delete this commit before merge.

This commit replaces the image's `src` by `data-original-src` to make
the `src` generation mechanism kick in when the involved blocks are
dragged and in case the theme is used by the website configurator.
(Also reduced image quality to stop runbot from complaining about
images > 100Kb)

task-2888667