[IMP] base,tools,mass_mailing,{test_}mail{_full}: turns attachments as links
The email was not sent when it was too heavy for the outgoing server. This
implementation solves the problem by allowing to specify (and even autodetect
when supported) a maximum size for email for each outgoing server. If that
limit is reached for an email, its attachment are automatically turned into
downloadable links inserted at the end of the email so that it can be sent and
the recipient can download the attachment through the provided links.

Note that there could still be an error if the recipient server cannot handle
the email size. But it can be mitigated by setting a low value for the max
email size of the outgoing servers.

Technical note: You can specify the maximum email size on each outgoing server
(ir.mail_server) and for the smtp server specified on the command line with
the added parameter max-outgoing-email-size (default 5MB).

Task-2474726