[IMP] api: deterministic Transaction.envs
Transaction.envs is a Weakset, making the iteration randomized
Iteration on weakset is used when trying to find an existing environment
or when determining the env to use to flush.

Bot example could lead to random errors as the one fixed by #29058

This commit proposes to replace Weakset with a WeakValueDictionary.
The iteration on WeakValueDictionary is the same as the insertion order
as for a dict.

```
    from weakref import WeakValueDictionary, WeakSet

    class X():
        def __init__(self, val):
            self.val = str(val)

        def __repr__(self):
            return self.val

        def __hash__(self):
            return hash(self.val)

    elems = [X(i) for i in ['a', '1', 'b', '2', 'c']]
    s = WeakSet(elems)
    d = WeakValueDictionary({elem.val:elem for elem in elems})
    print('WeakSet ', list(s))
    print('WeakDict', list(d.values()))
    del elems
    print('WeakSet ', list(s))
    print('WeakDict', list(d.values()))

```
This should also make the matching of an existing env faster if there is
many of .