[IMP] digest: fix outlook desktop design
Currently, the digest email template looks different in the Windows Outlook
desktop app. This issue is generated because it does not support margins and
padding in div and many modern CSS properties.

This commit fixes the issue by using HTML tables, inline CSS in such cases, and
old CSS properties to solve the above problem.

task-2717426