[IMP] mail: improves link redirection in emails by restoring the menu
It improves link redirection in emails by restoring the menu, easing the user
to take action. For example, when clicking on "view task" in an email related
to a project, it restores the Project menu.

Technical note:
- when redirecting to the page to view a model from a link in a message, we try
to find the best menu root to display to the user based on the
ir.actions.act_window bound to menu items visible to the user and matching the
model we want to display.
- finding the best root menu for a model should take 3 queries as asserted in
the tests if the cache is big enough (the search is done only on the menu
visible by the user):
-- to fetch the groups of the user in the decorator of _visible_menu_ids
-- to fetch the menu of the user
-- to fetch the act_window contained in the user's menu

- We have choosen:
-- to implements an heuristic rather than defining the root menu in each model
because it would have mean a lot of changes and a sort of manual heuristic
anyway
-- to find the root menu at the click rather than at the URL generation to
spread the performance cost
-- to not cache the mapping of the model to menu id to spare memory and as it
shouldn't impact the performance too much with simple queries (small added load
when the user access the URL)

Task-2834445