[IMP] mail, test_{crm_full|mail}: improves notification
It improves the notification of created item by adding a notification body to
ease the identification of the item created. For example, when creating a task,
the project name is included in the notification body.

Technical note:
- as mail_thread takes the message to be sent to followers, the body message is
common for all recipients. And as there might be multiple causes why a specific
recipient is notified, we had to make the message generic. That's why the
message can include multiple reasons not tailored to each recipient.
- _get_auto_subscription_subtypes is used to determine parent model because it
is cached even if we only need a small portion of its return.
- to determine the parent model instance, getattr on thread is used instead of
val_list because the logic to link the parent model instance to the created
instance is handled by the model class (often by setting values from the
context and not from val_list).
- Fetching the parent model increases a bit the number of queries in tests.

Task-2834445