[REF] web: use FileInput and remove unnecessary components
During the Owl conversion of fields, two components
were introduced to handle uploading and downloading
files. But it already existed a FileInput component
which had a similar purpose and the downloader
component could simply be removed since it's
basic purpose can be placed directly in the template.

This commit adapts those components to use FileInput:
BinaryField, ImageField and PdfViewerField