[FIX] {website_}event_booth_sale: fix booth price list computation
The price list rules were applied only when choosing the booth in the front end
but not in the card nor in the sale order in the backend. This solves the
problem by using the same technic introduced for event ticket.

Technical notes:
- this fix relies on the mechanism that allows to override the product price
with the booth price using a context variable.
- the context variable record_being_sold refers here to a record
event_booth_category and not an event_booth because the price reduce is
computed on event_booth_category and not on event_booth. As multiple
event_booth can be linked to an event_booth_category, we cannot know in
an event_booth_category which one of the event_booth is being sold.
- unfortunately, sale_order_line._get_product_price_context cannot be
overridden here to simply supply the booths in record_being_sold through the
context and then doing the sum in product._get_base_or_overridden_price because
in this case it doesn't work with pricelist rule "fixed" price; in that case
you get the same unit price no matter the number of booth ordered in the same
sale order line (as a reminder, a sale order line can contains multiple booth
while having quantity set to 1).

Co-authored-by: Florian Charlier <flch@odoo.com>
Co-authored-by: Pierre-Yves Dufays <pydu@odoo.com>

Task-2802757