[IMP] product: set up a mechanism to override the product price
Allow other modules to override the price during the computation of the final
price (based on the rules of the price list, ...).

This feature is useful in event module where we need to compute the final price
for different tickets with different price but linked to the same product.
The same applies for event booth.

Technical note: the price override relies on
- the context variable "record_being_sold" that gives the id of the record
being sold
- the detailed_type of the product that identify the record model on which the
price is queried through a mapping (DETAILED_TYPE_MODELS)
See example in event and event_booth.

Co-authored-by: Florian Charlier <flch@odoo.com>
Co-authored-by: Pierre-Yves Dufays <pydu@odoo.com>

Task-2802757