[FIX] mail, sms: excludes sms_blacklist/optout/duplicate in the chatter
Duplicate, opted out, blacklisted sms message were displayed in the chatter
suggesting to the user that it was sent while it was not. This solves the
problem by filtering out the sms that were not sent because duplicated,
blacklisted or opted out.

Technical note:
- in order to identify such mail.message, mail.message are linked to sms.sms so
that we can get the failure type from a mail.message
- the controller that returns the message for the chatter now execute 2 queries
(one in sudo and an other one without sudo) which cause the added query in some
tests

Task-2834862