[FIX] website: prevent dropping generated ToC outside its snippet
Both the generated entries of the table of content and the paragraphs
content of the table of content can be dropped outside of the table of
content snippet.

This commit prevents this by removing the move operations from the table
of content's generated navigation menu and content wrapper.
The Position option of the navigation menu can still be used to move it
before, on top, or after the content wrapper.

task-2752391