[REF] core: reimplement trigger trees from the field dependency graph
The registry now uses a "TriggerGraph" to represent the dependencies of
all its fields.  The construction of the graph is faster than building
all trigger trees.  The trigger tree of any subset of fields is now made
by the graph itself.

This provides a much better tradeoff for building trigger trees.
Building a trigger tree might be a bit slower than before, because it is
always built from the trigger graph, instead of merging and pruning
existing trigger trees.  However, the first trigger tree only requires
to build the trigger graph (which is roughly linear in the total number
of dependencies), while before it required to build all trigger trees.