[IMP] link_tracker,mass_mailing: make different trackers from labels
Purpose: Independently track clicks resolving to the same url from
different links on the same mailing if associated with a different label.

Note that as we add a criterion in the unicity constraint, we are adding a
dimension that wasn't used before. Therefore, no problem should arise in
existing databases.

Also, the pattern matching is a bit complex to support images because of
* The optional use of an 'alt' attribute
* alterations performed to properly format the message's content for various
email clients.

Ill-formed urls are updated in the mailing templates.

Tests are included and their formatting is condensed and uniformized.

Task-2788965