[IMP] web: move back the frontend scrolling out of the #wrapwrap
This finally reverts the most annoying part of [1]: the main scrolling
is now back to be the natural browser behavior instead of being moved to
the #wrapwrap element.

As a reminder, this was done for the 14.0 version and above since we did
not want a double scrollbar on the right of the website builder (and
some other advantages at the time). With the recent "website-in-backend"
work done at [2], the website is now displayed inside an iframe, and the
editor is outside of it. There is thus no reason to have the scrollbar
moved out its natural position.

Having it at its natural position fixes some behaviors we were not able
to fix in stable 14.0 and 15.0, notably:

- Many mobile browsers (including Chrome on Android) were not able to
  understand that the #wrapwrap had the main scrollbar. We were able to
  scroll of course, but some mobile behaviors were lost:

    - Reaching the top again (by scrolling to the top) and continuing
      scrolling with your finger was not reloading the page.

    - The mobile browser UI was not being hidden when scrolling down.

- Some browser extensions were not working as were based on scrolling
  the natural scrollbar.

This could also result in simplifying the code (many things were made
generic so the features would work wherever the main scrollbar is). This
commit although keeps that generic code for now. It probably make sense
to get rid of it in the future though as we are always bug-fixing one
mode without bug-fixing the other.

Note: this also fixes some code that relied on the fact #wrapwrap was
the one scrolling instead of using Odoo utils.

[1]: https://github.com/odoo/odoo/commit/4e7be69825163c0a0ff41c882a196fc7f3158fb3
[2]: https://github.com/odoo/odoo/commit/31cc10b91dc7762e23b4bde9b945be0c4ce3fe3b