[IMP] survey: improve mcq design
Currently, in the survey for single/multiple choice questions, the layout is
not consistent. For example, if there are four choices available, the
layout is 3-1 (3 small options in the first row, 1 full-width option in the last
row). Apart from that, if the string for options is long, the radio/check
box comes at the bottom right corner, which should always be at the
top right corner.

This commit does the following improvements:

1/ Improve the choices layout to make it beautiful. now there will be only one
   answer per row similar to the 'One Page Per Question' layout.
2/ The radio / check-box icon will be displayed in the top right corner
   regardless of the length of the option strings.
3/ If there is one or more image(s) in the answers then we will follow 2x2
   layout instead of one row per answer.

TaskID-2712296