[IMP] web_editor, website: adds label input in the editor panel
This commit adds a "text" input below the URL input in the link editor
so user can use that zone when they need to edit link name.

This url name is updated only when the input is changed so we don't
loose formatting (e.g. if one letter is bold).

Only the formatting applied to the entire selection is kept. (e.g. if
you have a "link <b>like</b> this", nothing will be in bold if you
update the label in the editor panel.

Also, if there is a media in the selection (image, icon), the
"text" input will be hidden.

task-2900529